Move Camera2D tests to their own file
